JIRA Issues from dumaguete tester
JIRA Issues from dumaguete tester

Fixed route activate account
Changed password Fixed Issues
Custom appbar
Fixed bug forgot password
 Fixed issues

+++ lib/http/http_request.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
 
+import 'package:dart_ping/dart_ping.dart';
 import 'package:http/http.dart' as http;
 import 'package:luvpark_get/http/api_keys.dart';
 
@@ -10,7 +11,21 @@ class HttpRequest {
 
   const HttpRequest({required this.api, this.parameters});
 
-  static Future<http.Response> fetchDataWithTimeout(link) async {
+  static Future<bool> checkInternetConn() async {
+    final ping = Ping('google.com', count: 5);
+    bool isConnected = false;
+
+    await for (var event in ping.stream) {
+      if (event.response != null) {
+        isConnected = true;
+        break;
+      }
+    }
+    return isConnected;
+  }
+
+  static Future<http.Response> fetchDataWithTimeout(
+      Future<http.Response> link) async {
     const timeoutDuration = Duration(seconds: 10);
 
     return await link.timeout(timeoutDuration, onTimeout: () {
@@ -20,6 +35,11 @@ class HttpRequest {
   }
 
   Future<dynamic> get() async {
+    bool isNetconn = await checkInternetConn();
+    if (!isNetconn) {
+      return "No Internet";
+    }
+
     var links = http.get(
         Uri.parse(Uri.decodeFull(Uri.https(ApiKeys.gApiURL, api).toString())),
         headers: {"Content-Type": 'application/json; charset=utf-8'});
@@ -38,6 +58,10 @@ class HttpRequest {
   }
 
   Future<dynamic> post() async {
+    bool isNetconn = await checkInternetConn();
+    if (!isNetconn) {
+      return "No Internet";
+    }
     var links = http.post(
         Uri.parse(Uri.decodeFull(Uri.https(ApiKeys.gApiURL, api).toString())),
         headers: {"Content-Type": 'application/json; charset=utf-8'},
@@ -56,6 +80,10 @@ class HttpRequest {
   }
 
   Future<dynamic> postBody() async {
+    bool isNetconn = await checkInternetConn();
+    if (!isNetconn) {
+      return "No Internet";
+    }
     var links = http.post(
         Uri.parse(Uri.decodeFull(Uri.https(ApiKeys.gApiURL, api).toString())),
         headers: {"Content-Type": 'application/json; charset=utf-8'},
@@ -74,6 +102,10 @@ class HttpRequest {
   }
 
   Future<dynamic> put() async {
+    bool isNetconn = await checkInternetConn();
+    if (!isNetconn) {
+      return "No Internet";
+    }
     var links = http.put(
         Uri.parse(Uri.decodeFull(Uri.https(ApiKeys.gApiURL, api).toString())),
         headers: {"Content-Type": "application/json"},
@@ -93,6 +125,10 @@ class HttpRequest {
   }
 
   Future<dynamic> deleteData() async {
+    bool isNetconn = await checkInternetConn();
+    if (!isNetconn) {
+      return "No Internet";
+    }
     var links = http.delete(
         Uri.parse(Uri.decodeFull(Uri.https(ApiKeys.gApiURL, api).toString())),
         headers: {"Content-Type": 'application/json; charset=utf-8'},
@@ -112,6 +148,10 @@ class HttpRequest {
   }
 
   Future<dynamic> linkToPage() async {
+    bool isNetconn = await checkInternetConn();
+    if (!isNetconn) {
+      return "No Internet";
+    }
     var links = http.get(Uri.https("luvpark.ph", "/terms-of-use"),
         headers: {"Content-Type": 'application/json; charset=utf-8'});
     try {
